added a voice button to the toolbar to switch to voice IME
