+++ app/src/main/res/layout/keyboard_view_keyboard.xml
@@ -74,12 +74,12 @@
         </com.simplemobiletools.keyboard.views.InlineContentViewHorizontalScrollView>
 
         <ImageView
-            android:id="@+id/voice_input"
+            android:id="@+id/voice_input_button"
             android:layout_width="@dimen/toolbar_icon_height"
             android:layout_height="@dimen/toolbar_icon_height"
             android:layout_marginEnd="@dimen/medium_margin"
             android:background="?android:attr/selectableItemBackgroundBorderless"
-            android:contentDescription="Voice Input"
+            android:contentDescription="@string/voice_input"
             android:padding="@dimen/small_margin"
             android:src="@drawable/ic_microphone_vector"
             app:layout_constraintBottom_toBottomOf="parent"
