+++ app/src/main/res/values-ckb/strings.xml
@@ -38,4 +38,5 @@
     <string name="start_sentences_capitalized">گەورەکردنی یەکەم پیتی لاتینی</string>
     <!-- Emojis -->
     <string name="emojis">خەندەکان</string>
+    <string name="voice_input">بگۆڕە بۆ شێوازی نوسينی دەنگ</string>
 </resources>
