+++ app/src/main/res/values-fr/strings.xml
@@ -38,6 +38,7 @@
     <string name="start_sentences_capitalized">Commencer les phrases par une majuscule</string>
     <!-- Emojis -->
     <string name="emojis">Émojis</string>
+    <string name="voice_input">Switch to voice input method</string>
     <!--
         Haven't found some strings? There's more at
         https://github.com/SimpleMobileTools/Simple-Commons/tree/master/commons/src/main/res
