+++ app/src/main/res/values-lv/strings.xml
@@ -38,6 +38,7 @@
     <string name="start_sentences_capitalized">Sākt teikumus ar lielo burtu</string>
     <!-- Emojis -->
     <string name="emojis">Emocijzīmes</string>
+    <string name="voice_input">Switch to voice input method</string>
     <!--
         Haven't found some strings? There's more at
         https://github.com/SimpleMobileTools/Simple-Commons/tree/master/commons/src/main/res
