+++ app/src/main/res/values-pa-rPK/strings.xml
@@ -38,4 +38,5 @@
     <string name="start_sentences_capitalized">Start sentences with a capital letter</string>
     <!-- Emojis -->
     <string name="emojis">ایموجیاں</string>
+    <string name="voice_input">Switch to voice input method</string>
 </resources>
