+++ app/src/main/res/values-sl/strings.xml
@@ -38,4 +38,5 @@
     <string name="start_sentences_capitalized">Start sentences with a capital letter</string>
     <!-- Emojis -->
     <string name="emojis">Emoji-ji</string>
+    <string name="voice_input">Switch to voice input method</string>
 </resources>
