+++ app/src/main/res/values-sv/strings.xml
@@ -38,6 +38,7 @@
     <string name="start_sentences_capitalized">Börja meningar med stor bokstav</string>
     <!-- Emojis -->
     <string name="emojis">Emojier</string>
+    <string name="voice_input">Switch to voice input method</string>
     <!--
         Haven't found some strings? There's more at
         https://github.com/SimpleMobileTools/Simple-Commons/tree/master/commons/src/main/res
