+++ app/src/main/res/values-zh-rCN/strings.xml
@@ -38,6 +38,7 @@
     <string name="start_sentences_capitalized">句子开头使用大写字母</string>
     <!-- Emojis -->
     <string name="emojis">表情符号</string>
+    <string name="voice_input">Switch to voice input method</string>
     <!--
         Haven't found some strings? There's more at
         https://github.com/SimpleMobileTools/Simple-Commons/tree/master/commons/src/main/res
